+++ examples/ccs/ccs_server.py
@@ -1,39 +1,28 @@
-from charm4py import charm, Chare, Array, Future, Reducer, Group, liveviz, coro
-import time
-import random
+from charm4py import charm, Chare, Array, Future, Reducer, Group
 
-class Unit(Chare):
-  
-  def __init__(self):
-    self.colors = [(200, 0, 0), (0, 200, 0), (0, 0, 200)]
-    
-  def reqImg(self):
-    for i in range(50):
-      self.particles = []
-      
-      for _ in range(300):
-        x = random.randint(0, 49)
-        y = random.randint(0, 49)
-        
-        color = random.choice(self.colors)
-        
-        self.particles.append((x, y, color))
-      
-      data = bytearray(50 * 50 * 3)
-      
-      for x, y, (r, g, b) in self.particles:
-        pixel_index = (y * 50 + x) * 3
-        data[pixel_index] = r
-        data[pixel_index + 1] = g
-        data[pixel_index + 2] = b
-      
-      liveviz.LiveViz.deposit(data, self, self.thisIndex[0]*50, self.thisIndex[1]*50, 50, 50, 800, 800)
+def handler(msg):
+    print("CCS Ping handler called on " + str(charm.myPe()))
+
+    msg = msg.rstrip('\x00') #removes null characters from the end
+    answer = "Hello to sender " + str(msg) + " from PE " + str(charm.myPe()) + ".\n"
+    # send the answer back to the client
+    charm.CcsSendReply(answer)
+
+class RegisterPerChare(Chare):
+
+    def register(self, return_future, handler):
+        charm.CcsRegisterHandler("ping2", handler)
+        charm.CcsRegisterHandler("ping", handler)
+        self.reduce(return_future, Reducer.nop)
 
 def main(args):
-    units = Array(Unit, dims=(16,16))
-    config = liveviz.Config()
-    liveviz.LiveViz.init(config, units, units.reqImg, poll=True)
-    units.reqImg()
+    # No need to initialize converse, because charm.start does this
+    # just register the handler
+    reg_wait = Future()
+    registers = Group(RegisterPerChare)
+    registers.register(reg_wait, handler)
+    reg_wait.get()
     print("CCS Handlers registered . Waiting for net requests...")
+
 
 charm.start(main)
